+++ common/src/main/java/net/conczin/mca/client/gui/VillagerEditorScreen.java
@@ -486,8 +486,7 @@ public class VillagerEditorScreen extends Screen implements SkinListUpdateListen
                         ProfessionsMCA.ADVENTURER,
                         ProfessionsMCA.CULTIST,
                 }) {
-                    MutableComponent text = Component.translatable("entity.minecraft.villager." + p);
-                    ButtonWidget widget = addRenderableWidget(new ButtonWidget(width / 2 + (right ? DATA_WIDTH / 2 : 0), y, DATA_WIDTH / 2, 20, text, b -> {
+                    ButtonWidget widget = addRenderableWidget(new ButtonWidget(width / 2 + (right ? DATA_WIDTH / 2 : 0), y, DATA_WIDTH / 2, 20, p.name(), b -> {
                         CompoundTag compound = new CompoundTag();
                         compound.putString("profession", BuiltInRegistries.VILLAGER_PROFESSION.getKey(p).toString());
                         syncVillagerData();